dashboard/backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
anndata>=0.10
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
scipy>=1.11
